Parsing BlE command accroding MotoconSDK
Description:
Parsing BLE command accroding the define of protobuf in motoconSDK

Testing and Verification Instruction:
NA

Type: Task
Issue: AUTOOTM-IXWW22-6244
Reviewed-on: https://gerrit.garmin.com.tw/code-review/c/autooem/products/nxp/rt1172/+/86736
Tested-by: Kuo, Dilbert <dilbert.kuo@garmin.com>
Approve-Change: Kuo, Dilbert <dilbert.kuo@garmin.com>

+++ osek-canbus/client_diag/inc/client_mem.h
@@ -50,7 +50,7 @@ extern client_ReturnType client_mem_storage_identifier_data
 
 extern void client_mem_storage_server_code
     (
-    uint16 server_code
+    uint32 server_code
     );
 
 boolean client_mem_get_storage_init_dtc_state
